Info. del Libro a FIREBASE

+++ app/src/main/java/mx/itesm/thegoldenbook/models/Album.kt
@@ -1,7 +1,7 @@
 package mx.itesm.thegoldenbook.models
 
 data class Album(
-    val albumId: Int,
+    val albumId: Long,
     val ownerId: String,
     val titulo: String,
     val rutaPortada: String,
